Update dioxus version of demo to 0.7.1
