+++ demo/.gitignore
@@ -1,12 +1,7 @@
 # Generated by Cargo
 # will have compiled files and executables
-/target/
-/dist/
-/static/
-/.dioxus/
-
-# this file will generate by tailwind:
-/assets/tailwind.css
+/target
+.DS_Store
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
