+++ demo/Cargo.toml
@@ -1,16 +1,21 @@
 [package]
 name = "demo"
 version = "0.1.0"
-authors = ["wakefullynx <mail@wakefullynx.dev>"]
+authors = ["philippebner <29914922+philippebner@users.noreply.github.com>"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-
-dioxus = { version = "0.5", features = ["web"] }
+dioxus = { version = "0.7.1", features = [] }
 smallest-enclosing-circle = { path = "../" }
-dioxus-free-icons = { version = "0.8", features = ["bootstrap"] }
+dioxus-free-icons = { version = "0.10.0", features = ["bootstrap"] }
 
-# Debug
-dioxus-logger = "0.5.1"
+[features]
+default = ["web"]
+# The feature that are only required for the web = ["dioxus/web"] build target should be optional and only enabled in the web = ["dioxus/web"] feature
+web = ["dioxus/web"]
+# The feature that are only required for the desktop = ["dioxus/desktop"] build target should be optional and only enabled in the desktop = ["dioxus/desktop"] feature
+desktop = ["dioxus/desktop"]
+# The feature that are only required for the mobile = ["dioxus/mobile"] build target should be optional and only enabled in the mobile = ["dioxus/mobile"] feature
+mobile = ["dioxus/mobile"]
