+++ demo/Dioxus.toml
@@ -1,41 +1,19 @@
 [application]
 
-# App (Project) Name
-name = "demo"
-
-# Dioxus App Default Platform
-# web, desktop, fullstack
-default_platform = "web"
-
-# `build` & `serve` dist path
-out_dir = "dist"
-
-# resource (assets) file folder
-asset_dir = "assets"
-
 [web.app]
 
 # HTML title tag content
-title = "Smallest enclosing circle demo"
+title = "Smallest Enclosing Circle Demo"
 
 #base_path = "smallest-enclosing-circle-demo"
-
-[web.watcher]
-
-# when watcher trigger, regenerate the `index.html`
-reload_html = true
-
-# which files or dirs will be watcher monitoring
-watch_path = ["src", "assets"]
 
 # include `assets` in web platform
 [web.resource]
 
-# CSS style file
+# Additional CSS style files
+style = []
 
-style = ["/tailwind.css"]
-
-# Javascript code file
+# Additional JavaScript files
 script = []
 
 [web.resource.dev]
