+++ demo/README.md
@@ -1,17 +1,29 @@
 # Development
 
-1. Install npm: https://docs.npmjs.com/downloading-and-installing-node-js-and-npm
-2. Install the tailwind css cli: https://tailwindcss.com/docs/installation
-3. Run the following command in the root of the project to start the tailwind CSS compiler:
+Your new jumpstart project includes basic organization with an organized `assets` folder and a `components` folder.
+If you chose to develop with the router feature, you will also have a `views` folder.
 
-```bash
-npx tailwindcss -i ./input.css -o ./assets/tailwind.css --watch
+```
+project/
+├─ assets/ # Any assets that are used by the app should be placed here
+├─ src/
+│  ├─ main.rs # The entrypoint for the app.
+│  ├─ components/
+│  │  ├─ mod.rs # Defines the components module
+│  │  ├─ hero.rs # The Hero component for use in the home page
+├─ Cargo.toml # The Cargo.toml file defines the dependencies and feature flags for your project
 ```
 
-Run the following command in the root of the project to start the Dioxus dev server:
+### Serving Your App
+
+Run the following command in the root of your project to start developing with the default platform:
 
 ```bash
-dx serve --hot-reload
+dx serve
 ```
 
-- Open the browser to http://localhost:8080
+To run for a different platform, use the `--platform platform` flag. E.g.
+```bash
+dx serve --platform desktop
+```
+
