+++ demo/src/components/mod.rs
@@ -1,0 +1,6 @@
+//! The components module contains all shared components for our app. Components are the building blocks of dioxus apps.
+//! They can be used to defined common UI elements like buttons, forms, and modals. In this template, we define a Hero
+//! component  to be used in our app.
+
+mod smallest_enclosing_circle_demo;
+pub use smallest_enclosing_circle_demo::SmallestEnclosingCircleDemo;
